refactor(angular): move non-delegate component providers to common
