+++ packages/angular/common/src/providers/toast-controller.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
-import { ToastOptions, toastController } from '@ionic/core';
+import type { ToastOptions } from '@ionic/core/components';
+import { toastController } from '@ionic/core/components';
 
-import { OverlayBaseController } from '../util/overlay';
+import { OverlayBaseController } from '../utils/overlay';
 
 @Injectable({
   providedIn: 'root',
